modules/github-runner/cloud-function/requirements.txt
google-cloud-compute==1.15.0
google-cloud-secret-manager==2.16.4
functions-framework==3.5.0
